Fixed damage bug and made weapon.
