+++ Such Life/UserSettings/EditorUserSettings.asset
@@ -24,10 +24,10 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c731104101d6a18212a3520280d73d6e5272cacf238e0f323
       flags: 0
     RecentlyUsedScenePath-6:
-      value: 22424703114646680e0b0227036c72111f192b292926237e38271427fb
+      value: 22424703114646680e0b0227036c5c1f1b1527272d2627372824183df6ae2136ebf32f
       flags: 0
     RecentlyUsedScenePath-7:
-      value: 22424703114646680e0b0227036c5c1f1b1527272d2627372824183df6ae2136ebf32f
+      value: 22424703114646680e0b0227036c72111f192b292926237e38271427fb
       flags: 0
     RecentlyUsedScenePath-8:
       value: 22424703114646680e0b0227036c5b111b161f2f3f2b233e2867083debf42d
